refactor: update all components on Homepage and update tests
