+++ react-components/src/components/SearchBar.tsx
@@ -1,4 +1,4 @@
-import React, { FormEventHandler, useRef } from 'react';
+import React, { useRef, useCallback } from 'react';
 import glass from '../assets/search_glass.png';
 import { useDispatch, useSelector } from 'react-redux';
 import { RootState } from '../redux/store';
@@ -6,25 +6,22 @@ import { setInputValue } from '../redux/features/searchSlice';
 
 export function SearchBar() {
   const searchRef = useRef<HTMLInputElement | null>(null);
-  const dispatch = useDispatch();
   const inputValue = useSelector((state: RootState) => state.search.inputValue);
+  const dispatch = useDispatch();
 
-  const handleSubmit: FormEventHandler<HTMLFormElement> = (event) => {
-    event.preventDefault();
-    if (searchRef.current) {
-      dispatch(setInputValue(searchRef.current.value));
-    }
-  };
+  const handleSubmit = useCallback(
+    (event: React.FormEvent<HTMLFormElement>) => {
+      event.preventDefault();
+      if (searchRef.current) {
+        dispatch(setInputValue(searchRef.current.value));
+      }
+    },
+    [dispatch]
+  );
 
   return (
     <form data-testid="searchbarForm" className="searchBar" onSubmit={handleSubmit}>
-      <input
-        data-testid="searchbar"
-        type="text"
-        placeholder="Search..."
-        defaultValue={inputValue || ''}
-        ref={searchRef}
-      />
+      <input data-testid="searchbar" type="text" defaultValue={inputValue || ''} ref={searchRef} />
       <img data-testid="search-image" className="searchGlass" src={glass} alt="search image" />
     </form>
   );
